refactor: migrate hooks dir to root
