+++ hooks/useDeviceType.ts
