+++ pages/biz/index.tsx
@@ -1,6 +1,6 @@
 import { Container, Typography } from '@mui/material';
 import bizLayout from '@/features/layout/biz/BizLayout';
-import { useAuth } from '@/hooks/useAuth';
+import { useAuth } from 'hooks/useAuth';
 import { Loading } from '@/components/asset/Loading';
 
 import { NextPageWithLayout } from '@/services/types/types';
