+++ pages/mypage/account/login-details.tsx
@@ -5,7 +5,7 @@ import {
   TextField,
   Typography,
 } from '@mui/material';
-import { useAuth } from '../../../services/hooks/useAuth';
+import { useAuth } from '../../../hooks/useAuth';
 import type { NextPage } from 'next';
 import { Loading } from '../../../components/asset/Loading';
