+++ pages/showcase/authentification/LoginCenteredExample.tsx
@@ -3,7 +3,7 @@ import type { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { BrandLogo } from '@/components/asset/BrandLogo';
-import { useAuth } from '@/services/hooks/useAuth';
+import { useAuth } from 'hooks/useAuth';
 import { Loading } from '@/components/asset/Loading';
 import BrothLink from '@/components/link/BrothLink';
 import { BrothTypography } from '@/components/typography/BrothTypography';
