+++ pages/showcase/authentification/LoginSocials.tsx
@@ -12,7 +12,7 @@ import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { BrandLogo } from '@/components/asset/BrandLogo';
 import { LoginForm } from '@/features/authorization/LoginForm';
-import { useAuth } from '@/services/hooks/useAuth';
+import { useAuth } from 'hooks/useAuth';
 import { Loading } from '@/components/asset/Loading';
 import BrothLink from '@/components/link/BrothLink';
 import { BrothTypography } from '@/components/typography/BrothTypography';
