+++ pages/showcase/forms/AddressFormExample.tsx
@@ -7,7 +7,7 @@ import DeleteIcon from '@mui/icons-material/Delete';
 import { useForm } from 'react-hook-form';
 import yup from '@/lib/yup';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { useSnackbar } from '@/hooks/useSnackbar';
+import { useSnackbar } from 'hooks/useSnackbar';
 import BrothSnackbar from '@/components/snackbar/BrothSnackbar';
 import { YubinBangoCore } from 'lib/yubinbango-core';
