+++ pages/showcase/forms/MultiStepFormExample.tsx
@@ -4,7 +4,7 @@ import { BrothTypography } from '@/components/typography/BrothTypography';
 import { Box, Container, Grid, Paper, Step } from '@mui/material';
 import { FormProvider, useForm, useFormContext } from 'react-hook-form';
 import yup from '@/lib/yup';
-import { useSnackbar } from '@/hooks/useSnackbar';
+import { useSnackbar } from 'hooks/useSnackbar';
 import BrothSnackbar from '@/components/snackbar/BrothSnackbar';
 import { BrothStepper } from '@/components/stepper/BrothStepper';
 import { BrothStepLabel } from '@/components/stepper/BrothStepLabel';
